add send_image_url. Move bot.Bot code to send_api.SendApiClient. update bot.Bot to maintain backwards compatibility.
